Custom User Field Permissions
For the initial release of the User type of dynamic field, the field should be forced to be invisible, not required, and not editable to end users.

For now, this is a limitation set for privacy reasons to ensure users do not indirectly have access to the User Directory. Additionally, if we did want to allow this field for Users, we would need to set up the ajax lookup on the client side.

The future plan would be to allow Users to see Users in their Organization only.