Combine Laravel and osTicket
This codebase has all of the files from osTicket as well as Laravel.

We need to make our current codebase work alongside Laravel (in-place/side-by-side migration). New features will be created in Laravel and we can gradually migrate old code. This allows us to not have to halt development of new features.