Feature: Google 2FA Plugin Extras
This commit adds code to osTicket that is necessary for a new plugin, Google Two Factor Authentication, to work. It also adds a new function, pluginExists, to the Plugin class so that we can reuse code if we need to show something in the helpdesk based on if a plugin exists or not.