Feature: IP Banlist/Whitelist
This feature adds options to ban IPs from creating Tickets or Users in the helpdesk. There is also an option to whitelist certain IPs so that the banlist rules will not apply to them.

This feature can be configured by going to:

Admin Panel | Settings | Tickets | IP Banlist Configuration

From here, the Admin has the option to:
1. Ban all blacklisted IPs from creating Tickets/Users
	- for this option, IPs are not automatically added to the blacklist
2. Limit the amount of Tickets that can be in an Open status for IPs
	- Once the max limit is hit, IPs are automatically added to this list
3. Disable this feature completely by setting the 'Maximum Open Tickets - IP' to 0
	- This is the default setting

IPs can be added or removed from the Banlist or Whitelist by going to:

Admin Panel | Manage | Lists

The 'Whitelisted IPs' and 'Banned IPs' lists are created automatically upon installing osTicket or by configuring the feature for existing installations. These lists are set up to where they cannot be deleted. Both lists will only accept valid IPs.

If an IP is deleted from the Banlist, it is automatically added to the Whitelist so that it no longer gets blocked from the helpdesk. IPs can be added/removed from the Whitelist like any other list in the system.